parser/requirements.txt
ttn
mysql.connector
flask
requests